% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_Normalized_Gradient.R
\name{compare_points}
\alias{compare_points}
\title{Compare Points Based on Preference Feedback}
\usage{
compare_points(x1, x2)
}
\arguments{
\item{x1}{A numeric vector representing the first point.}

\item{x2}{A numeric vector representing the second point.}
}
\value{
An integer, +1 if \code{f(x1) < f(x2)}, -1 otherwise.
}
\description{
A placeholder function that simulates comparison feedback between two points.
Returns +1 if \code{f(x1) < f(x2)}, -1 otherwise.
}
