% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DuelingOptimization-package.R
\docType{package}
\name{DuelingOptimization-package}
\alias{DuelingOptimization}
\alias{DuelingOptimization-package}
\title{DuelingOptimization: Convex Optimization with Preference (Dueling) Feedback}
\description{
The goal is to find the optimal point with the least possible query complexity, without the luxury of even zeroth-order feedback.
}
\author{
\strong{Maintainer}: Arpan Ghosh \email{arpan2001@tamu.edu}

}
\keyword{internal}
