% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_Normalized_Gradient.R
\name{f_value}
\alias{f_value}
\title{Calculate the Function Value at a Point}
\usage{
f_value(x)
}
\arguments{
\item{x}{A numeric vector representing the point at which to evaluate \code{f}.}
}
\value{
A numeric value representing the function value at \code{x}.
}
\description{
A simple function to calculate the value of \code{f} at a given point. This is used
for tracking purposes in the \code{beta_NGD} algorithm.
}
