% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sign_recovery_wrapper.R
\name{signRecovery}
\alias{signRecovery}
\title{Perform Sign Recovery with a Objective Function}
\usage{
signRecovery(x, y, delta, f)
}
\arguments{
\item{x}{A numeric vector representing the first point.}

\item{y}{A numeric vector representing the second point.}

\item{delta}{A positive numeric value specifying the confidence parameter.}

\item{f}{A function representing the objective function. This function
must take a numeric vector as input and return a scalar numeric value.}
}
\value{
An integer: +1 if the recovered sign is positive, -1 otherwise.
}
\description{
It performs sign recovery by repeatedly comparing two points using the provided objective function.
}
\examples{
# Define an objective function
f <- function(x) sum(x^3-x)
# Perform sign recovery
signRecovery(x = c(1, 2), y = c(3, 4), delta = 0.01, f = f)
}
