% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_Normalized_Gradient.R
\name{compare_points}
\alias{compare_points}
\title{Compare Two Points Based on Objective Function}
\usage{
compare_points(x1, x2, f)
}
\arguments{
\item{x1}{A numeric vector representing the first point.}

\item{x2}{A numeric vector representing the second point.}

\item{f}{The objective function to be minimized. It must take a numeric vector as input and return a numeric value.}
}
\value{
An integer, +1 if \code{f(x1) < f(x2)}, -1 otherwise.
}
\description{
Compares two points based on the values of a given objective function.
The function returns +1 if the objective function value at the first point
is smaller than the value at the second point, and -1 otherwise.
}
\keyword{internal}
